Functional on switch with some caveats. Fixed inverse R stick Y axis.
Switch requires masquerading as a supported controller type, unfortunately. This means that the Teensy can no longer be identified as a Teensy board to work on the Switch. The current idea is to have the Teensy identify itself as both a Switch-compatible controller and as a PC Joystick. However, this means that PCs actually detect 2 controllers being connected, though only the Joystick interface actually communicates with the PC, as the other interface is specifically for the Switch protocol. It may be possible to switch the VID/PID of the board dynamically, but I haven't yet figured out a way to do that. Still investigating...